Remove persistent history tree
It might have something to do with the weird segfaults i'm having due to
some cache or saved file or such. And i often had issues trying to
recover said history. And it doesn't save it too long anyway, given that
i like to put it in tmp, so i haven't had any use for it.