Remove undo files etcetera in vim
Since i don't do any longer work in vim it's not necessary and just
pollutes tmp. Also i still do :w every third second anyway. undofile
is a bit of a loss in the rare situations that i use it, but with the
emacs support for the same being so-so i should definitely get into a
habit of preemptively backing up stuff i don't want to break anyway.